Revert "Adding Orders Feature"
This reverts commit ba5c9a5874f60937215168c5e33f6b76d30991ff.
