+++ src/components/controller/Products/index.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import { toggleShowDescription, placeOrder, clearOrderResponses, clearProductQuantities, updateQuantity } from 'actions';
+import { openModal, toggleShowDescription, updateQuantity } from 'actions';
 import Products from 'components/view/Products';
 
 const mapStateToProps = (state) => {
@@ -7,12 +7,10 @@ const mapStateToProps = (state) => {
     return {
         products: state.products,
         userPermissions: state.user.permissions,
-        asyncResponses: state.asyncResponses,
-        user: state.user,
     };
 };
 
-const mapDispatchToProps = { toggleShowDescription, placeOrder, clearOrderResponses, clearProductQuantities, updateQuantity };
+const mapDispatchToProps = { openModal, toggleShowDescription, updateQuantity };
 
 const ProductsController = connect(
     mapStateToProps, mapDispatchToProps
