+++ src/components/view/Admin/index.js
@@ -1,13 +1,8 @@
 import React from 'react';
-import Orders from 'components/controller/Orders';
 
 const AdminPage = props => {
-
     return (
-        <div className="page">
-            <h2>Admin View</h2>
-            <Orders {...props} />
-        </div>
+        <div className="page">ADMIN VIEW</div>
     );
 };
