--- src/components/view/Orders/index.js
@@ -1,7 +1,0 @@
-import Orders from './orders.js';
-import OrderDetails from './order-details.js';
-
-export {
-    Orders,
-    OrderDetails
-};
