+++ src/components/view/Products/index.js
@@ -1,23 +1,18 @@
 import React from 'react';
 import { Redirect } from 'react-router-dom';
-import { Cart } from 'components/view/Common/Cart';
+import Cart from 'components/view/Common/Cart';
 import NotAuthorized from 'components/view/NotAuthorized';
 
 
 const Products = (props) => {
     
-    const { user, location, userPermissions, products, toggleShowDescription, placeOrder, updateQuantity, clearOrderResponses, clearProductQuantities, asyncResponses } = props;
+    const { location, userPermissions, openModal, products, toggleShowDescription, updateQuantity } = props;
 
     const cartProps = {
         products,
         toggleShowDescription,
         updateQuantity,
-        rowIndexExpanded: 0,
-        placeOrder,
-        asyncResponses,
-        clearOrderResponses,
-        clearProductQuantities,
-        user,
+        rowIndexExpanded: 0
     };
     
     const notAuthProps = {
