+++ src/constants.js
@@ -13,23 +13,16 @@ export const LOGIN_GOOGLE_REQUEST = 'LOGIN_GOOGLE_REQUEST';
 export const OPEN_MODAL = 'OPEN_MODAL';
 export const CLOSE_MODAL = 'CLOSE_MODAL';
 
+/** USER PERMISSION STATUSES **/
+// export const ANONYMOUS = 'ANONYMOUS';
+// export const PENDING = 'PENDING';
+// export const CUSTOMER = 'CUSTOMER';
+// export const SELLER = 'SELLER';
+// export const ADMIN = 'ADMIN';
+
 /** PRODUCTS */
 export const GET_PRODUCTS = 'GET_PRODUCTS';
 export const UPDATE_PRODUCTS = 'UPDATE_PRODUCTS';
-export const PLACE_ORDER = 'PLACE_ORDER';
-export const CLEAR_PRODUCT_QUANTITIES = 'CLEAR_PRODUCT_QUANTITIES';
-
-/** ORDERS */
-export const GET_ORDERS = 'GET_ORDERS';
-export const UPDATE_ORDERS = 'UPDATE_ORDERS';
-export const TOGGLE_ORDER_DETAILS = 'TOGGLE_ORDER_DETAILS';
-
-/* ASYNC */
-export const RESPONSE_CODE_SUCCESS = 200;
-export const RESPONSE_CODE_FAIL = 400;
-export const PLACE_ORDER_RESPONSE = 'PLACE_ORDER_RESPONSE';
-export const CLEAR_ORDER_RESPONSES = 'CLEAR_ORDER_RESPONSES';
-
 
 /** ROW EXPANSION **/
 export const TOGGLE_SHOW_DESCRIPTION = 'TOGGLE_SHOW_DESCRIPTION';
