+++ src/reducers/index.js
@@ -1,8 +1,6 @@
 import modal from './modal.js';
 import user from './user.js';
 import products from './products.js';
-import orders from './orders.js';
-import asyncResponses from './async-responses.js';
 
 function isLoggedIn(state = false, action) {
     return state;
@@ -13,8 +11,6 @@ const reducers = {
     modal,
     user,
     products,
-    orders,
-    asyncResponses,
 };
 
 export default reducers;
