+++ src/reducers/products.js
@@ -2,8 +2,6 @@ import {
     TOGGLE_SHOW_DESCRIPTION,
     UPDATE_QUANTITY,
     UPDATE_PRODUCTS,
-    PLACE_ORDER,
-    CLEAR_PRODUCT_QUANTITIES,
 } from '../constants';
 
 const initialState = [
@@ -41,7 +39,7 @@ const initialState = [
 ];
 
 function products(state = initialState, action) {
-    const { type, row, productId, optionKey, quantity, products, orderMeta, response } = action;
+    const { type, row, productId, optionKey, quantity, products } = action;
 
     if (type === TOGGLE_SHOW_DESCRIPTION) {
 
@@ -56,7 +54,7 @@ function products(state = initialState, action) {
         state[state.findIndex(product => { return product.item === productId})].options
             .forEach(option => {
                 if (option.key === optionKey) {
-                    option.quantity = parseInt(quantity,10);
+                    option.quantity = quantity;
                 }
                 return option;
             });
@@ -68,18 +66,6 @@ function products(state = initialState, action) {
         return [
             ...products,
         ];
-    } else if (type === CLEAR_PRODUCT_QUANTITIES) {
-        return state.map(product => {
-            if (product.options) {
-                product.options = product.options.map( option => {
-                    option.quantity = 0;
-                    return option;
-                });
-                return product;
-            } else {
-                return product;
-            }
-        });
     } else {
         return state;
     }
