+++ src/sagas/index.js
@@ -2,7 +2,6 @@ import { fork } from 'redux-saga/effects';
 import authentication from './authentication';
 import user from './user';
 import product from './product';
-import order from './order';
 
 /**
  * use one root saga to yield all other side effect sagas
@@ -12,7 +11,6 @@ function* sagas() {
         fork(authentication),
         fork(user),
         fork(product),
-        fork(order),
     ];
 }
