+++ src/sagas/user.js
@@ -32,7 +32,6 @@ function* updateUser({userData}) {
             uid: userData.uid,
             displayName: userData.displayName,
             permissions: userData.permissions,
-            email: userData.email,
         };
     } else {
         cleanUser = {...currentUser, ...userData};
